pyproject.toml
[tool.poetry]
name = "ibgateway"
version = "0.1.0"
description = "IB Gateway Docker Image with noVNC and Screenshot Support"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "ibgateway"}]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = ">=1.0.0"
Pillow = ">=10.0.0"
numpy = {version = ">=1.24.0", optional = true}

[tool.poetry.extras]
full = ["numpy"]

[tool.poetry.scripts]
ibgateway-cli = "ibgateway.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.2.0,<9.0.0)"
]
